Moved set address functionality to base node interface

+++ Implementation/Power LoRa/Device/RadioDevice.cs
@@ -9,7 +9,7 @@
         #endregion
 
         #region Constructors
-        public RadioDevice(int address) : base()
+        public RadioDevice(byte address) : base()
 		{
 			Address = address;
 			Connected = false;
